fix/change auth folder structure
change auth api folder structure with create components folder and then create services,controllers,validations,custom-validations and routes file in it and also change env.js file name to config.js and change paths of all modules which is modified and also change all routes url
