+++ .env.example
@@ -5,4 +5,3 @@ ACCESS_TOKEN_KEY =
 REFRESH_TOKEN_KEY = 
 ACCESS_TOKEN_EXPIRY = 
 REFRESH_TOKEN_EXPIRY = 
-NEW_KEY =
