+++ package.json
@@ -12,7 +12,7 @@
   "description": "",
   "dependencies": {
     "cookie-parser": "^1.4.7",
-    "dotenv": "^16.4.5",
+    "dotenv": "^16.4.7",
     "dotenv-safe": "^9.1.0",
     "express": "^4.21.1",
     "express-session": "^1.18.1",
