+++ schemas/userMdl.js
@@ -2,18 +2,18 @@ import mongoose from "mongoose";
 
 const userSchema = new mongoose.Schema(
   {
-    name: { type: String, required: true },
+    name: { type: String, required: [true, "Name is required (from DB)"] },
     email: {
       type: String,
-      required: true,
+      required: [true, "Email ID is required"],
       match: [
         /^[\w-]+(\.[\w-]+)*@([\w-]+\.)+[a-zA-Z]{2,7}$/,
-        "Please provide a valid email address",
+        "Please provide a valid email address (from DB)",
       ],
     },
     password: {
       type: String,
-      required: true,
+      required: [true, "Password is required (from DB)"],
     },
     registeredAt: { type: Date },
     refreshToken: { type: String },
